move document creation logic into adapter
